requirements.txt
multiformats
mysql-connector-python
pathspec
rich
pydantic
python-dotenv
# Testing and Development
pytest
pytest-cov
pytest-mock
pytest-timeout
pyyaml
nltk
# Pre-commit hooks
pre-commit
